split interface.h
