+++ src/recover.h
@@ -1,0 +1,11 @@
+#ifndef __RECOVER_H__
+#define __RECOVER_H__
+
+int FTI_CheckFile(char *fn, long fs, char* checksum);
+int FTI_CheckErasures(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
+        FTIT_topology* FTI_Topo, FTIT_checkpoint* FTI_Ckpt,
+        int *erased);
+int FTI_RecoverFiles(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
+        FTIT_topology* FTI_Topo, FTIT_checkpoint* FTI_Ckpt);
+
+#endif // __RECOVER_H__
