+++ src/tools.c
@@ -446,769 +446,6 @@ void FTI_FreeTypesAndGroups(FTIT_execution* FTI_Exec)
   free(FTI_Exec->H5groups);
 }
 
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      It creates h5datatype (hid_t) from definitions in FTI_Types
-  @param      ftiType        FTI_Type type
-
-  This function creates (opens) hdf5 compound type. Should be called only
-  before saving checkpoint in HDF5 format. Build-in FTI's types are always open.
-
- **/
-/*-------------------------------------------------------------------------*/
-void FTI_CreateComplexType(FTIT_type* ftiType, FTIT_type** FTI_Type)
-{
-  char str[FTI_BUFS];
-  if (ftiType->h5datatype > -1) {
-    //This type already created
-    sprintf(str, "Type [%d] is already created.", ftiType->id);
-    FTI_Print(str, FTI_DBUG);
-    return;
-  }
-
-  if (ftiType->structure == NULL) {
-    //Save as array of bytes
-    sprintf(str, "Creating type [%d] as array of bytes.", ftiType->id);
-    FTI_Print(str, FTI_DBUG);
-    ftiType->h5datatype = H5Tcopy(H5T_NATIVE_CHAR);
-    H5Tset_size(ftiType->h5datatype, ftiType->size);
-    return;
-  }
-
-  hid_t partTypes[FTI_BUFS];
-  int i;
-  //for each field create and rank-dimension array if needed
-  for (i = 0; i < ftiType->structure->length; i++) {
-    sprintf(str, "Type [%d] trying to create new type [%d].", ftiType->id, ftiType->structure->field[i].typeID);
-    FTI_Print(str, FTI_DBUG);
-    FTI_CreateComplexType(FTI_Type[ftiType->structure->field[i].typeID], FTI_Type);
-    partTypes[i] = FTI_Type[ftiType->structure->field[i].typeID]->h5datatype;
-    if (ftiType->structure->field[i].rank > 1) {
-      //need to create rank-dimension array type
-      hsize_t dims[FTI_BUFS];
-      int j;
-      for (j = 0; j < ftiType->structure->field[i].rank; j++) {
-        dims[j] = ftiType->structure->field[i].dimLength[j];
-      }
-      sprintf(str, "Type [%d] trying to create %d-D array of type [%d].", ftiType->id, ftiType->structure->field[i].rank, ftiType->structure->field[i].typeID);
-      FTI_Print(str, FTI_DBUG);
-      partTypes[i] = H5Tarray_create(FTI_Type[ftiType->structure->field[i].typeID]->h5datatype, ftiType->structure->field[i].rank, dims);
-    } else {
-      if (ftiType->structure->field[i].dimLength[0] > 1) {
-        //need to create 1-dimension array type
-        sprintf(str, "Type [%d] trying to create 1-D [%d] array of type [%d].", ftiType->id, ftiType->structure->field[i].dimLength[0], ftiType->structure->field[i].typeID);
-        FTI_Print(str, FTI_DBUG);
-        hsize_t dim = ftiType->structure->field[i].dimLength[0];
-        partTypes[i] = H5Tarray_create(FTI_Type[ftiType->structure->field[i].typeID]->h5datatype, 1, &dim);
-      }
-    }
-  }
-
-  //create new HDF5 datatype
-  sprintf(str, "Creating type [%d].", ftiType->id);
-  FTI_Print(str, FTI_DBUG);
-  ftiType->h5datatype = H5Tcreate(H5T_COMPOUND, ftiType->size);
-  sprintf(str, "Type [%d] has hid_t %ld.", ftiType->id, (long)ftiType->h5datatype);
-  FTI_Print(str, FTI_DBUG);
-  if (ftiType->h5datatype < 0) {
-    FTI_Print("FTI failed to create HDF5 type.", FTI_WARN);
-  }
-
-  //inserting fields into the new type
-  for (i = 0; i < ftiType->structure->length; i++) {
-    sprintf(str, "Insering type [%d] into new type [%d].", ftiType->structure->field[i].typeID, ftiType->id);
-    FTI_Print(str, FTI_DBUG);
-    herr_t res = H5Tinsert(ftiType->h5datatype, ftiType->structure->field[i].name, ftiType->structure->field[i].offset, partTypes[i]);
-    if (res < 0) {
-      FTI_Print("FTI faied to insert type in complex type.", FTI_WARN);
-    }
-  }
-
-}
-#endif
-
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      It closes h5datatype
-  @param      ftiType        FTI_Type type
-
-  This function destroys (closes) hdf5 compound type. Should be called
-  after saving checkpoint in HDF5 format.
-
- **/
-/*-------------------------------------------------------------------------*/
-void FTI_CloseComplexType(FTIT_type* ftiType, FTIT_type** FTI_Type)
-{
-  char str[FTI_BUFS];
-  if (ftiType->h5datatype == -1 || ftiType->id < 11) {
-    //This type already closed or build-in type
-    sprintf(str, "Cannot close type [%d]. Build in or already closed.", ftiType->id);
-    FTI_Print(str, FTI_DBUG);
-    return;
-  }
-
-  if (ftiType->structure != NULL) {
-    //array of bytes don't have structure
-    int i;
-    //close each field
-    for (i = 0; i < ftiType->structure->length; i++) {
-      sprintf(str, "Closing type [%d] of compound type [%d].", ftiType->structure->field[i].typeID, ftiType->id);
-      FTI_Print(str, FTI_DBUG);
-      FTI_CloseComplexType(FTI_Type[ftiType->structure->field[i].typeID], FTI_Type);
-    }
-  }
-
-  //close HDF5 datatype
-  sprintf(str, "Closing type [%d].", ftiType->id);
-  FTI_Print(str, FTI_DBUG);
-  herr_t res = H5Tclose(ftiType->h5datatype);
-  if (res < 0) {
-    FTI_Print("FTI failed to close HDF5 type.", FTI_WARN);
-  }
-  ftiType->h5datatype = -1;
-}
-#endif
-
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      It creates a group and all it's children
-  @param      ftiGroup        FTI_H5Group to be create
-  @param      parentGroup     hid_t of the parent
-
-  This function creates hdf5 group and all it's children. Should be
-  called only before saving checkpoint in HDF5 format.
-
- **/
-/*-------------------------------------------------------------------------*/
-void FTI_CreateGroup(FTIT_H5Group* ftiGroup, hid_t parentGroup, FTIT_H5Group** FTI_Group)
-{
-  ftiGroup->h5groupID = H5Gcreate2(parentGroup, ftiGroup->name, H5P_DEFAULT, H5P_DEFAULT, H5P_DEFAULT);
-  if (ftiGroup->h5groupID < 0) {
-    FTI_Print("FTI failed to create HDF5 group.", FTI_WARN);
-    return;
-  }
-
-  int i;
-  for (i = 0; i < ftiGroup->childrenNo; i++) {
-    FTI_CreateGroup(FTI_Group[ftiGroup->childrenID[i]], ftiGroup->h5groupID, FTI_Group); //Try to create the child
-  }
-}
-#endif
-
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      It opens a group and all it's children
-  @param      ftiGroup        FTI_H5Group to be opened
-  @param      parentGroup     hid_t of the parent
-
-  This function opens hdf5 group and all it's children. Should be
-  called only before recovery in HDF5 format.
-
- **/
-/*-------------------------------------------------------------------------*/
-void FTI_OpenGroup(FTIT_H5Group* ftiGroup, hid_t parentGroup, FTIT_H5Group** FTI_Group)
-{
-  ftiGroup->h5groupID = H5Gopen2(parentGroup, ftiGroup->name, H5P_DEFAULT);
-  if (ftiGroup->h5groupID < 0) {
-    FTI_Print("FTI failed to open HDF5 group.", FTI_WARN);
-    return;
-  }
-
-  int i;
-  for (i = 0; i < ftiGroup->childrenNo; i++) {
-    FTI_OpenGroup(FTI_Group[ftiGroup->childrenID[i]], ftiGroup->h5groupID, FTI_Group); //Try to open the child
-  }
-}
-#endif
-
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      It closes a group and all it's children
-  @param      ftiGroup        FTI_H5Group to be closed
-
-  This function closes (destoys) hdf5 group and all it's children. Should be
-  called only after saving checkpoint in HDF5 format.
-
- **/
-/*-------------------------------------------------------------------------*/
-void FTI_CloseGroup(FTIT_H5Group* ftiGroup, FTIT_H5Group** FTI_Group)
-{
-  char str[FTI_BUFS];
-  if (ftiGroup->h5groupID == -1) {
-    //This group already closed, in tree this is error
-    snprintf(str, FTI_BUFS, "Group %s is already closed?", ftiGroup->name);
-    FTI_Print(str, FTI_WARN);
-    return;
-  }
-
-  int i;
-  for (i = 0; i < ftiGroup->childrenNo; i++) {
-    FTI_CloseGroup(FTI_Group[ftiGroup->childrenID[i]], FTI_Group); //Try to close the child
-  }
-
-  herr_t res = H5Gclose(ftiGroup->h5groupID);
-  if (res < 0) {
-    FTI_Print("FTI failed to close HDF5 group.", FTI_WARN);
-  }
-  ftiGroup->h5groupID = -1;
-}
-#endif
-
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      It creates the global dataset in the VPR file.
-  @param      FTI_Exec        Execution metadata.
-  @return     integer         FTI_SCES if successful.
-
-  Creates global dataset (shared among all ranks) in VPR file. The dataset
-  position will be the group assigned to it by calling the FTI API function 
-  'FTI_DefineGlobalDataset'.
- **/
-/*-------------------------------------------------------------------------*/
-int FTI_CreateGlobalDatasets( FTIT_execution* FTI_Exec )
-{
-    
-    FTIT_globalDataset* dataset = FTI_Exec->globalDatasets;
-    while( dataset ) {
-
-        // create file space
-        dataset->fileSpace = H5Screate_simple( dataset->rank, dataset->dimension, NULL );
-
-        if(dataset->fileSpace < 0) {
-            char errstr[FTI_BUFS];
-            snprintf( errstr, FTI_BUFS, "Unable to create space for dataset #%d", dataset->id );
-            FTI_Print(errstr,FTI_EROR);
-            return FTI_NSCS;
-        }
-
-        // create dataset
-        hid_t loc = dataset->location->h5groupID;
-        hid_t tid = FTI_Exec->FTI_Type[dataset->type.id]->h5datatype;
-        dataset->hdf5TypeId = tid;
-        hid_t fsid = dataset->fileSpace;
-        
-        // FLETCHER CHECKSUM NOT SUPPORTED FOR PARALLEL I/O IN HDF5
-        //hid_t dcpl = H5Pcreate (H5P_DATASET_CREATE);
-        //H5Pset_fletcher32 (dcpl);
-        //
-        //hsize_t *chunk = malloc( sizeof(hsize_t) * dataset->rank );
-        //chunk[0] = chunk[1] = 4096;
-        //H5Pset_chunk (dcpl, 2, chunk);
-
-        dataset->hid = H5Dcreate( loc, dataset->name, tid, fsid, H5P_DEFAULT, H5P_DEFAULT, H5P_DEFAULT );
-        if(dataset->hid < 0) {
-            char errstr[FTI_BUFS];
-            snprintf( errstr, FTI_BUFS, "Unable to create dataset #%d", dataset->id );
-            FTI_Print(errstr,FTI_EROR);
-            return FTI_NSCS;
-        }
-
-        dataset->initialized = true;
-
-        dataset = dataset->next;
-
-    }
-
-    return FTI_SCES;
-
-}
-#endif
-
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      Opens global dataset in VPR file.
-  @param      FTI_Exec        Execution metadata.
-  @return     integer         FTI_SCES if successful.
-
-  This function is the analog to 'FTI_CreateGlobalDatasets' on recovery.
- **/
-/*-------------------------------------------------------------------------*/
-int FTI_OpenGlobalDatasets( FTIT_execution* FTI_Exec )
-{
-    
-    hsize_t *dims = NULL; 
-    hsize_t *maxDims = NULL;
-
-    char errstr[FTI_BUFS];
-    FTIT_globalDataset* dataset = FTI_Exec->globalDatasets;
-    while( dataset ) {
-
-        dims = (hsize_t*) realloc( dims, sizeof(hsize_t)*dataset->rank ); 
-        maxDims = (hsize_t*) realloc( maxDims, sizeof(hsize_t)*dataset->rank );
-        
-        // open dataset
-        hid_t loc = dataset->location->h5groupID;
-        hid_t tid = FTI_Exec->FTI_Type[dataset->type.id]->h5datatype;
-        dataset->hdf5TypeId = tid;
-
-        dataset->hid = H5Dopen( loc, dataset->name, H5P_DEFAULT );
-        if(dataset->hid < 0) {
-            snprintf( errstr, FTI_BUFS, "failed to open dataset '%s'", dataset->name );
-            FTI_Print( errstr, FTI_WARN );
-            return FTI_NSCS;
-        }
-        
-        // get file space and check if rank and dimension coincide for file and execution
-        hid_t fsid = H5Dget_space( dataset->hid );
-        if( fsid > 0 ) {
-            int rank = H5Sget_simple_extent_ndims( fsid );
-            if( rank == dataset->rank ) {
-                H5Sget_simple_extent_dims( fsid, dims, maxDims );
-                if( memcmp( dims, dataset->dimension, sizeof(hsize_t)*rank ) ) {
-                    snprintf( errstr, FTI_BUFS, "stored and requested dimensions of dataset '%s' differ!", dataset->name );
-                    FTI_Print( errstr, FTI_WARN );
-                    return FTI_NSCS;
-                }
-            } else {
-                snprintf( errstr, FTI_BUFS, "stored and requested rank of dataset '%s' differ (stored:%d != requested:%d)!", dataset->name, rank, dataset->rank );
-                FTI_Print( errstr, FTI_WARN );
-                return FTI_NSCS;
-            }
-        } else {
-            snprintf( errstr, FTI_BUFS, "failed to acquire data space information of dataset '%s'", dataset->name );
-            FTI_Print( errstr, FTI_WARN );
-            return FTI_NSCS;
-        }    
-        dataset->fileSpace = fsid;
-        
-        dataset->initialized = true;
-
-        dataset = dataset->next;
-
-    }
-
-    if( dims ) { free( dims ); }
-    if( maxDims ) { free( maxDims ); }
-
-    return FTI_SCES;
-
-}
-#endif
-
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      Closes global datasets in VPR file 
-  @param      FTI_Exec        Execution metadata.
-  @return     integer         FTI_SCES if successful.
- **/
-/*-------------------------------------------------------------------------*/
-int FTI_CloseGlobalDatasets( FTIT_execution* FTI_Exec )
-{
-    
-    FTIT_globalDataset* dataset = FTI_Exec->globalDatasets;
-    while( dataset ) {
-
-        H5Sclose(dataset->fileSpace);
-
-        H5Dclose(dataset->hid);
-
-        dataset = dataset->next;
-
-    }
-
-    return FTI_SCES;
-
-}
-#endif
-
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      Writes global dataset subsets into VPR file.
-  @param      FTI_Data        Dataset metadata.
-  @return     integer         FTI_SCES if successful.
- **/
-/*-------------------------------------------------------------------------*/
-herr_t FTI_WriteSharedFileData( FTIT_dataset FTI_Data )
-{
-
-    if( FTI_Data.sharedData.dataset ) {
-        
-        // hdf5 datatype
-        hid_t tid = FTI_Data.sharedData.dataset->hdf5TypeId;
-
-        // dataset hdf5-id
-        hid_t did = FTI_Data.sharedData.dataset->hid;
-
-        // shared dataset file space
-        hid_t fsid = FTI_Data.sharedData.dataset->fileSpace;
-
-        // shared dataset rank
-        int ndim = FTI_Data.sharedData.dataset->rank;
-
-        // shared dataset array of nummber of elements in each dimension
-        hsize_t *count = FTI_Data.sharedData.count;
-
-        // shared dataset array of the offsets for each dimension
-        hsize_t *offset = FTI_Data.sharedData.offset;
-
-        // create dataspace for subset of shared dataset
-        hid_t msid = H5Screate_simple( ndim, count, NULL );
-        if(msid < 0) {
-            char errstr[FTI_BUFS];
-            snprintf( errstr, FTI_BUFS, "Unable to create space for var-id %d in dataset #%d", FTI_Data.id, FTI_Data.sharedData.dataset->id );
-            FTI_Print(errstr,FTI_EROR);
-            return FTI_NSCS;
-        }
-        // select range in shared dataset in file
-        if( H5Sselect_hyperslab(fsid, H5S_SELECT_SET, offset, NULL, count, NULL) < 0 ) {
-            char errstr[FTI_BUFS];
-            snprintf( errstr, FTI_BUFS, "Unable to select sub-space for var-id %d in dataset #%d", FTI_Data.id, FTI_Data.sharedData.dataset->id );
-            FTI_Print(errstr,FTI_EROR);
-            return FTI_NSCS;
-        }
-
-        // enable collective buffering
-        hid_t plid = H5Pcreate( H5P_DATASET_XFER );
-        H5Pset_dxpl_mpio(plid, H5FD_MPIO_COLLECTIVE);
-
-        // write data in file
-        if( H5Dwrite(did, tid, msid, fsid, plid, FTI_Data.ptr) < 0 ) {
-            char errstr[FTI_BUFS];
-            snprintf( errstr, FTI_BUFS, "Unable to write var-id %d of dataset #%d", FTI_Data.id, FTI_Data.sharedData.dataset->id );
-            FTI_Print(errstr,FTI_EROR);
-            return FTI_NSCS;
-        }
-
-        H5Sclose( msid );
-        H5Pclose( plid );
-
-    }
-
-    return FTI_SCES;
-
-}
-#endif
-
-#ifdef ENABLE_HDF5 
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      Checks for matching dimension sizes of sub-sets
-  @param      FTI_Data        Dataset metadata.
-  @param      FTI_Exec        Execution metadata.
-  @return     integer         FTI_SCES if successful.
-
-  This function counts the number of elements of all sub-sets contained in a 
-  particular global dataset and accumulates to a total value. If the accu-
-  mulated value matches the number of elements defined for the global data-
-  set, FTI_SCES is returned. This function is called before the checkpoint
-  and before the recovery.
-
-  @todo it would be great to check for region overlapping too.
-
- **/
-/*-------------------------------------------------------------------------*/
-int FTI_CheckDimensions( FTIT_dataset * FTI_Data, FTIT_execution * FTI_Exec ) 
-{   
-
-    // NOTE checking for overlap is complicated and likely expensive
-    // since it requires sorting within all contributing processes.
-    // Thus, we check currently only the number of elements.
-    FTIT_globalDataset* dataset = FTI_Exec->globalDatasets;
-    while( dataset ) {
-        int i,j;
-        // sum of local elements
-        hsize_t numElemLocal = 0, numElemGlobal;
-        for( i=0; i<dataset->numSubSets; ++i ) {
-            hsize_t numElemSubSet = 1;
-            for( j=0; j<dataset->rank; j++ ) {
-                numElemSubSet *= FTI_Data[dataset->varIdx[i]].sharedData.count[j];
-            }
-            numElemLocal += numElemSubSet;
-        }
-        // number of elements in global dataset
-        hsize_t numElem = 1;
-        for( i=0; i<dataset->rank; ++i ) {
-            numElem *= dataset->dimension[i];
-        }
-        MPI_Allreduce( &numElemLocal, &numElemGlobal, 1, MPI_UNSIGNED_LONG_LONG, MPI_SUM, FTI_COMM_WORLD );
-        if( numElem != numElemGlobal ) {
-            char errstr[FTI_BUFS];
-            snprintf( errstr, FTI_BUFS, "Number of elements of subsets (accumulated) do not match number of elements defined for global dataset #%d!", dataset->id ); 
-            FTI_Print( errstr, FTI_WARN);
-            return FTI_NSCS;
-        }
-        dataset = dataset->next;
-    }
-    return FTI_SCES;
-}
-#endif
-
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      Checks if VPR file on restart
-  @param      FTI_Conf        Configuration metadata.
-  @param      ckptID          Checkpoint ID.
-  @return     integer         FTI_SCES if successful.
-
-  Checks if restart is possible for VPR file. 
-  1) Checks if file exist for prefix and directory, defined in config file.
-  2) Checks if is regular file.
-  3) Checks if groups and datasets can be accessed and if datasets can be 
-  read
-
-  If file found and sane, ckptID is set and FTI_SCES is returned.
-
- **/
-/*-------------------------------------------------------------------------*/
-int FTI_H5CheckSingleFile( FTIT_configuration* FTI_Conf, int *ckptID ) 
-{
-    char errstr[FTI_BUFS];
-    char fn[FTI_BUFS];
-    int res = FTI_SCES;
-    struct stat st;
-   
-    struct dirent *entry;
-    DIR *dir = opendir( FTI_Conf->h5SingleFileDir );
-
-    if ( dir == NULL ) {
-        snprintf( errstr, FTI_BUFS, "VPR directory '%s' could not be accessed.", FTI_Conf->h5SingleFileDir );
-        FTI_Print(errstr, FTI_EROR);
-        errno = 0;
-        return FTI_NSCS;
-    }
-
-    *ckptID = -1;
-
-    bool found = false;
-    while((entry = readdir(dir)) != NULL) {   
-        if(strcmp(entry->d_name,".") && strcmp(entry->d_name,"..")) {
-            int len = strlen( entry->d_name ); 
-            if( len > 14 ) {
-                char fileRoot[FTI_BUFS];
-                bzero( fileRoot, FTI_BUFS );
-                memcpy( fileRoot, entry->d_name, len - 14 );
-                char fileRootExpected[FTI_BUFS];
-                snprintf( fileRootExpected, FTI_BUFS, "%s", FTI_Conf->h5SingleFilePrefix );
-                if( strncmp( fileRootExpected, fileRoot, FTI_BUFS ) == 0 ) {
-                    int id_tmp;
-                    sscanf( entry->d_name + len - 14 + 3, "%08d.h5", &id_tmp );
-                    if( id_tmp > *ckptID ) {
-                        *ckptID = id_tmp;
-                        snprintf( fn, FTI_BUFS, "%s/%s-ID%08d.h5", FTI_Conf->h5SingleFileDir, FTI_Conf->h5SingleFilePrefix, *ckptID ); 
-                    }
-                    found = true;
-                }
-            }
-        }
-    }
-
-    if (!found) {
-        snprintf( errstr, FTI_BUFS, "unable to find matching VPR file (filename pattern: '%s-ID########.h5')!", FTI_Conf->h5SingleFilePrefix );
-        FTI_Print( errstr, FTI_WARN );
-        return FTI_NSCS;
-    }
-    
-    stat( fn, &st );
-    if( S_ISREG( st.st_mode ) ) {
-        hid_t fid = H5Fopen( fn, H5F_ACC_RDONLY, H5P_DEFAULT );
-        if( fid > 0 ) {
-            hid_t gid = H5Gopen1( fid, "/" );
-            if( gid > 0 ) {
-                res += FTI_ScanGroup( gid, fn );
-                H5Gclose(gid);
-            } else {
-                snprintf( errstr, FTI_BUFS, "failed to access root group in file '%s'", fn );
-                FTI_Print( errstr, FTI_WARN );
-                res = FTI_NSCS;
-            }
-            H5Fclose(fid);
-        } else {
-            snprintf( errstr, FTI_BUFS, "failed to open file '%s'", fn );
-            FTI_Print( errstr, FTI_WARN );
-            res = FTI_NSCS;
-        }
-    } else {
-        snprintf( errstr, FTI_BUFS, "'%s', is not a regular file!", fn );
-        FTI_Print( errstr, FTI_WARN );
-        res = FTI_NSCS;
-    }
-    return res;
-}
-#endif
-
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      Checks groups and datasets (callable recursively) 
-  @param      gid             Parent group ID.
-  @param      fn              File name.
-  @return     integer         FTI_SCES if successful.
-
-  This function analyses the group structure recursivley starting at group
-  'gid'. It steps down in sub groups and checks datasets of consistency.
-  The consistency check is performed if the dataset was created with the 
-  fletcher32 filter activated. A dataset read will return a negative value 
-  in that case if dataset corrupted.
-
- **/
-/*-------------------------------------------------------------------------*/
-int FTI_ScanGroup( hid_t gid, char* fn ) 
-{
-    int res = FTI_SCES;
-    char errstr[FTI_BUFS];
-    hsize_t nobj;
-    if( H5Gget_num_objs( gid, &nobj ) >= 0 ) {
-        int i;
-        for(i=0; i<nobj; i++) {
-            int objtype;
-            char dname[FTI_BUFS];
-            char gname[FTI_BUFS];
-            // determine if element is group or dataset
-            objtype = H5Gget_objtype_by_idx(gid, (size_t)i );
-            if( objtype == H5G_DATASET ) {
-                H5Gget_objname_by_idx(gid, (hsize_t)i, dname, (size_t) FTI_BUFS); 
-                // open dataset
-                hid_t did = H5Dopen1( gid, dname );
-                if( did > 0 ) {
-                    hid_t sid = H5Dget_space(did);
-                    hid_t tid = H5Dget_type(did);
-                    int drank = H5Sget_simple_extent_ndims( sid );
-                    size_t typeSize = H5Tget_size( tid );
-                    hsize_t *count = (hsize_t*) calloc( drank, sizeof(hsize_t) );
-                    hsize_t *offset = (hsize_t*) calloc( drank, sizeof(hsize_t) );
-                    count[0] = 1;
-                    char* buffer = (char*) malloc( typeSize );
-                    hid_t msid = H5Screate_simple( drank, count, NULL );
-                    H5Sselect_hyperslab(sid, H5S_SELECT_SET, offset, NULL, count, NULL);
-                    // read element to trigger checksum comparison
-                    herr_t status = H5Dread(did, tid, msid, sid, H5P_DEFAULT, buffer);
-                    if( status < 0 ) {
-                        snprintf( errstr, FTI_BUFS, "unable to read from dataset '%s' in file '%s'!", dname, fn );
-                        FTI_Print( errstr, FTI_WARN );
-                        res += FTI_NSCS;
-                    }
-                    H5Dclose(did);
-                    H5Sclose(msid);
-                    H5Sclose(sid);
-                    H5Tclose(tid);
-                    free( count );
-                    free( offset );
-                    free( buffer );
-                } else {
-                    snprintf( errstr, FTI_BUFS, "failed to open dataset '%s' in file '%s'", dname, fn );
-                    FTI_Print( errstr, FTI_WARN );
-                    res += FTI_NSCS;
-                }
-            }
-            // step down other group
-            if( objtype == H5G_GROUP ) {
-                H5Gget_objname_by_idx(gid, (hsize_t)i, gname, (size_t) FTI_BUFS); 
-                hid_t sgid = H5Gopen1( gid, gname );
-                if( sgid > 0 ) {
-                    res += FTI_ScanGroup( sgid, fn );
-                    H5Gclose(sgid);
-                } else {
-                    snprintf( errstr, FTI_BUFS, "failed to open group '%s' in file '%s'", gname, fn );
-                    FTI_Print( errstr, FTI_WARN );
-                    res += FTI_NSCS;
-                }
-            }
-        }
-    } else {
-        snprintf( errstr, FTI_BUFS, "failed to get number of elements in file '%s'", fn );
-        FTI_Print( errstr, FTI_WARN );
-        res += FTI_NSCS;
-    }
-    return FTI_SCES;
-}
-#endif
-
-#ifdef ENABLE_HDF5
-/*-------------------------------------------------------------------------*/
-/**
-  @brief      Reads a sub-set of a global dataset on recovery.
-  @param      FTI_Data        Dataset metadata.
-  @return     integer         FTI_SCES if successful.
- **/
-/*-------------------------------------------------------------------------*/
-herr_t FTI_ReadSharedFileData( FTIT_dataset FTI_Data )
-{
-
-    // hdf5 datatype
-    hid_t tid = FTI_Data.sharedData.dataset->hdf5TypeId;
-    
-    // dataset hdf5-id
-    hid_t did = FTI_Data.sharedData.dataset->hid;
-
-    // shared dataset file space
-    hid_t fsid = FTI_Data.sharedData.dataset->fileSpace;
-    
-    // shared dataset rank
-    int ndim = FTI_Data.sharedData.dataset->rank;
-
-    // shared dataset array of nummber of elements in each dimension
-    hsize_t *count = FTI_Data.sharedData.count;
-
-    // shared dataset array of the offsets for each dimension
-    hsize_t *offset = FTI_Data.sharedData.offset;
-
-    // create dataspace for subset of shared dataset
-    hid_t msid = H5Screate_simple( ndim, count, NULL );
-    if(msid < 0) {
-        char errstr[FTI_BUFS];
-        snprintf( errstr, FTI_BUFS, "Unable to create space for var-id %d in dataset #%d", FTI_Data.id, FTI_Data.sharedData.dataset->id );
-        FTI_Print(errstr,FTI_EROR);
-        return FTI_NSCS;
-    }
-
-    // select range in shared dataset in file
-    H5Sselect_hyperslab(fsid, H5S_SELECT_SET, offset, NULL, count, NULL);
-
-    // enable collective buffering
-    hid_t plid = H5Pcreate( H5P_DATASET_XFER );
-    H5Pset_dxpl_mpio(plid, H5FD_MPIO_COLLECTIVE);
-
-    // write data in file
-    herr_t status = H5Dread(did, tid, msid, fsid, plid, FTI_Data.ptr);
-    if(status < 0) {
-        char errstr[FTI_BUFS];
-        snprintf( errstr, FTI_BUFS, "Unable to read var-id %d from dataset #%d", FTI_Data.id, FTI_Data.sharedData.dataset->id );
-        FTI_Print(errstr,FTI_EROR);
-        return FTI_NSCS;
-    }
-
-    H5Sclose( msid );
-    H5Pclose( plid );
-
-    return status;
-
-}
-#endif
-
-#ifdef ENABLE_HDF5 
-void FTI_FreeVPRMem( FTIT_execution* FTI_Exec, FTIT_dataset* FTI_Data ) 
-{
-    FTIT_globalDataset * dataset = FTI_Exec->globalDatasets;
-    while( dataset ) {
-        if( dataset->dimension ) { free( dataset->dimension ); }
-        if( dataset->varIdx ) { free( dataset->varIdx ); }
-        FTIT_globalDataset * curr = dataset;
-        dataset = dataset->next;
-        free( curr );
-    }
-
-    int i=0;
-    for( ; i<FTI_Exec->nbVar; i++ ) {
-        if( FTI_Data[i].sharedData.offset ) {
-            free( FTI_Data[i].sharedData.offset );
-        }
-        if( FTI_Data[i].sharedData.count ) {
-            free( FTI_Data[i].sharedData.count );
-        }
-    }
-}
-#endif
-
 /*-------------------------------------------------------------------------*/
 /**
   @brief      It creates the basic datatypes and the dataset array.
@@ -1411,3 +648,4 @@ char* FTI_GetHashHexStr( const unsigned char* hash, int digestWidth, char* hashH
 
     return hashHexStr;
 }
+
