+++ src/topo.h
@@ -1,0 +1,13 @@
+#ifndef __TOPO_H__
+#define __TOPO_H__
+int FTI_SaveTopo(FTIT_configuration* FTI_Conf, FTIT_topology* FTI_Topo, char *nameList);
+int FTI_ReorderNodes(FTIT_configuration* FTI_Conf, FTIT_topology* FTI_Topo,
+        int *nodeList, char *nameList);
+int FTI_BuildNodeList(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
+        FTIT_topology* FTI_Topo, int *nodeList, char *nameList);
+int FTI_CreateComms(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
+        FTIT_topology* FTI_Topo, int *userProcList,
+        int *distProcList, int* nodeList);
+int FTI_Topology(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
+        FTIT_topology* FTI_Topo);
+#endif // __TOPO_H__
